% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Branding.R
\name{applyCRUKBrand}
\alias{applyCRUKBrand}
\title{Apply brand to plot}
\usage{
applyCRUKBrand(plot)
}
\arguments{
\item{plot}{the plot}
}
\description{
This function takes a plot and applies the CRUK brand and standard footnote text
}
\examples{
g <- applyCRUKBrand(plot)
grid.draw(g)
}
\keyword{CRUK}

