% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Branding.R
\name{getCRUKLogo}
\alias{getCRUKLogo}
\title{Get CRUK Logo Graphical Objects}
\usage{
getCRUKLogo()
}
\value{
The CRUK logo graphical object
}
\description{
This function returns the CRUK Logo at 25% of the png (bottom right quarter)
}
\examples{
getCRUKLogo()
}
\keyword{CRUK}

