% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Branding.R
\name{getFootnotesGrob}
\alias{getFootnotesGrob}
\title{Get Footnotes Graphical Object}
\usage{
getFootnotesGrob(footNoteText = NULL)
}
\arguments{
\item{footNoteText}{If NA then will use default text}
}
\value{
The footnotes graphical object
}
\description{
This function returns the footnotes graphical object
}
\examples{
getFootnotesGrob()
getFootnotesGrob("my footnote text")
}
\keyword{CRUK}

