% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Branding.R
\name{getFootnotesText}
\alias{getFootnotesText}
\title{Get Footnotes Text}
\usage{
getFootnotesText()
}
\value{
The footnotes text
}
\description{
This function returns the footnotes text
}
\examples{
getFootnotesText()
}
\keyword{CRUK}

